requirements.txt
litellm>=1.0.0
prompt-declaration-language>=0.1.0
PyYAML>=6.0